Honestly got no idea whats going on its been months since I wrote this
